move mobile uri to separate file
