+++ internal/driver/gomobile/file_android.go
@@ -12,33 +12,17 @@ char* readStream(uintptr_t jni_env, uintptr_t ctx, void* stream, int len, int* t
 void* saveStream(uintptr_t jni_env, uintptr_t ctx, char* uriCstr);
 void writeStream(uintptr_t jni_env, uintptr_t ctx, void* stream, char* data, int len);
 void closeStream(uintptr_t jni_env, uintptr_t ctx, void* stream);
-char* contentURIGetFileName(uintptr_t jni_env, uintptr_t ctx, char* uriCstr);
 */
 import "C"
 import (
 	"errors"
 	"io"
 	"os"
-	"path/filepath"
 	"unsafe"
 
 	"fyne.io/fyne/v2/storage/repository"
 	"github.com/fyne-io/mobile/app"
 )
-
-func (m mobileURI) Name() string {
-	if m.Scheme() == "content" {
-		result := contentURIGetFileName(m.systemURI)
-		if result != "" {
-			return result
-		}
-	}
-	return m.URI.Name()
-}
-
-func (m mobileURI) Extension() string {
-	return filepath.Ext(m.Name())
-}
 
 type javaStream struct {
 	stream unsafe.Pointer // java.io.InputStream
@@ -93,25 +77,6 @@ func openStream(uri string) unsafe.Pointer {
 		return nil
 	})
 	return stream
-}
-
-func contentURIGetFileName(uri string) string {
-	uriStr := C.CString(uri)
-	defer C.free(unsafe.Pointer(uriStr))
-
-	var filename string
-	app.RunOnJVM(func(_, env, ctx uintptr) error {
-		fnamePtr := C.contentURIGetFileName(C.uintptr_t(env), C.uintptr_t(ctx), uriStr)
-		vPtr := unsafe.Pointer(fnamePtr)
-		if vPtr == C.NULL {
-			return nil
-		}
-		filename = C.GoString(fnamePtr)
-		C.free(vPtr)
-
-		return nil
-	})
-	return filename
 }
 
 func nativeFileOpen(f *fileOpen) (io.ReadCloser, error) {
